izmene u api kontrolerima i api rutama

+++ app/Http/Controllers/API/LearnerController.php
@@ -6,6 +6,7 @@ namespace App\Http\Controllers\API;
 use App\Models\Learner;
 use Illuminate\Http\Request;
 use Illuminate\Routing\Controller;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Validation\Rules\Enum;
 
 class LearnerController extends Controller
@@ -28,7 +29,7 @@ class LearnerController extends Controller
             'first_name' => 'required|string|max:255',
             'last_name' => 'required|string|max:255',
             'email' => 'required|string|email|max:255|unique:users',
-            'password' => 'required|min:8',
+            'password' => 'required',
             'status' => [new Enum(ServerStatus::class)],
         ]);
 
@@ -48,9 +49,23 @@ class LearnerController extends Controller
      */
     public function update(Request $request, string $id)
     {
+      
+    if(Auth::getDefaultDriver() == 'admin'){
+
         $learner = Learner::find($id);
         $learner->update($request->all());
         return $learner;
+
+    }elseif(Auth::getDefaultDriver() == 'learners'){
+         if($id == Auth::guard('learners')->user()->id){
+        $learner = Learner::find($id);
+        $learner->update($request->all());
+        return $learner;
+         }else{
+            return "Unauthorized.";
+         }
+    }
+   
     }
 
     /**
@@ -58,7 +73,20 @@ class LearnerController extends Controller
      */
     public function destroy(string $id)
     {
-        return Learner::destroy($id);
+        if(Auth::getDefaultDriver() == 'admin'){
+
+            Learner::destroy($id);
+            return response()->json('Learner deleted.');
+    
+        }elseif(Auth::getDefaultDriver() == 'learners'){
+             if($id == Auth::guard('learners')->user()->id){
+                Learner::destroy($id);
+                return response()->json('Learner deleted.');
+             }else{
+                return "Unauthorized.";
+             }
+        }
+       
     }
 
     public function search($name)
